feat(zhtp): add GET /api/v1/blockchain/utxo/{tx_hash}/{output_index}/merkle-proof endpoint
Exposes the persistent UTXO Merkle tree to wallets so they can
query inclusion proofs for real ZK transaction proofs.

- Added  JSON type.
- Added  in .
- Wired route into handler dispatch.
- Added  test.
- Fixed  to expose  and dummy stubs
  when  is disabled, preventing workspace build breaks.
- Fixed pre-existing / mismatch in identity backup_recovery tests.

Part of Epic F wallet/client integration.

+++ zhtp/src/api/handlers/identity/backup_recovery.rs
@@ -2433,7 +2433,7 @@ mod tests {
             Option<String>,
             Vec<u8>,
             u64,
-            u64,
+            u128,
         )> = Vec::new();
         {
             let old = manager
